+++ packages/cli/examples/coralnet.ts
@@ -1,20 +1,20 @@
 interface Options {
   readonly httpUrl: string;
   readonly bech32prefix: string;
-  readonly hdPath: readonly Slip10RawIndex[];
+  readonly hdPath: HdPath;
   readonly gasPrice: GasPrice;
   readonly gasLimits: Partial<GasLimits<CosmWasmFeeTable>>; // only set the ones you want to override
 }
 
 const coralnetOptions: Options = {
-  httpUrl: 'https://lcd.coralnet.cosmwasm.com',
+  httpUrl: "https://lcd.coralnet.cosmwasm.com",
   gasPrice: GasPrice.fromString("0.025ushell"),
-  bech32prefix: 'coral',
+  bech32prefix: "coral",
   hdPath: makeCosmoshubPath(0),
-  gasLimits:  {
+  gasLimits: {
     upload: 1500000,
-  }
-}
+  },
+};
 
 const wallet = await Secp256k1Wallet.generate(12, coralnetOptions.hdPath, coralnetOptions.bech32prefix);
 const [{ address }] = await wallet.getAccounts();