+++ packages/launchpad-ledger/types/index.d.ts
@@ -1,0 +1,2 @@
+export { LaunchpadLedger } from "./launchpadledger";
+export { LedgerSigner } from "./ledgersigner";